Images sizes rearranged
So it can be seen in the zoom of 100%. Had to change the text a bit too as well (almost no change here throughout the text though). Most important than everything, font size of the text has been reduced from 12pt to 10pt. There's no other change available we can make in order to keep this in 8 pages for sure. If this version doesn't make it, we will have to start to cut off less important section/subsections.